+++ CHANGELOG.md
@@ -57,8 +57,10 @@
 - [11438](https://github.com/vegaprotocol/vega/issues/11438) - Add missing assignment to epoch to and from in `gameTeamScores` resolvers. 
 - [11442](https://github.com/vegaprotocol/vega/issues/11442) - Add validation to `submitAMM` to ensure the curves can be successfully generated. 
 - [11448](https://github.com/vegaprotocol/vega/issues/11448) - Fix team game score query with epoch filter. 
+- [11457](https://github.com/vegaprotocol/vega/issues/11457) - Fix cursor column ordering for game scores. 
 - [11454](https://github.com/vegaprotocol/vega/issues/11454) - Ensure ended transfers proper handling. 
   
+
 ## 0.76.1
 
 ### 🐛 Fixes